fixed doubles

+++ db/migrate/20150623060317_add_weight_to_goals.rb
@@ -1,5 +1,5 @@
 class AddWeightToGoals < ActiveRecord::Migration
   def change
-    add_column :goals, :weight, :double
+    add_column :goals, :weight, :float
   end
 end
